Struct dynamo::ioreg::builder::Builder

pub struct Builder<'a> { /* fields omitted */ }

Builds the AST from the parsed ioreg!() macro.

Methods

impl<'a> Builder<'a>

fn new(ioreg: IoRegInfo, parser: Parser, verbose: bool) -> Builder

Consumes the parsed IoRegInfo as well as the parser (for later syntax error placement).

fn build(&self) -> Vec<P<Item>>

Outputs Rust AST items representing the generated ioreg.

fn build_read_register<T>(&self, addr: u32, sp: Span) -> P<Expr> where T: ToAstType<T>

Builds a volatile_read() call that reads the entire mcu-proper-register (internally called an Segment).

For setters that are not fully-byte-aligned, a read is done before the write and the current value is masked and the re-written with the new bits properly set.

fn build_volatile_store_base<T>(&self,
addr: u32,
sp: Span)
-> ExprCallArgsBuilder<StmtExprBuilder<Identity>> where T: ToAstType<T>

Builds the skeleton of a volatile_store() call, with only the destination location set.

fn build_volatile_store<T>(&self,
addr: u32,
fn_def: &IoRegFuncDef,
seg: &IoRegSegmentInfo,
v: &FunctionValueType)
-> Stmt where T: ToAstType<T> + Narrow<u32>

Builds the volatile_store(dst, src, len) statement for a setter.

fn get_uint_const_val<T>(&self,
v: &FunctionValueType,
seg: &IoRegSegmentInfo)
-> T where T: ToAstType<T> + Narrow<u32>

Reads into the defined constants for this ioreg and returns the numeric literal behind the constant.

If not found as a defined internal-constant, a syntax error is placed an 0u32 is returned.

fn optimize_write<T>(&self,
fn_def: &IoRegFuncDef,
fn_prev: BlockBuilder<ExprBuilder<BlockBuilder<ItemImplMethodBuilder<ItemImplBuilder<Identity>>>>>,
off: &IoRegOffsetIndexInfo,
seg: &IoRegSegmentInfo)
-> BlockBuilder<ExprBuilder<BlockBuilder<ItemImplMethodBuilder<ItemImplBuilder<Identity>>>>> where T: ToAstType<T> + Narrow<u32> + BitAnd<T>

Considers the register, offset, and idex to be written to. Then optimizes and reads/writes needed to achieve the action.

If writing to a partial register, add a statement that says let fetched = volatile_load(addr as *const T). Then, we will mask this value and then do a volatile_store(addr as *mut T, masked | val)

TODO: This function is a little heavy duty to be done in a loop -- same responsibilities, but perhaps pass in the mask and address so their "expensive" calculations can be reduced

fn build_setter<T>(&self,
fn_def: &IoRegFuncDef,
seg: &IoRegSegmentInfo,
off: &IoRegOffsetIndexInfo,
fn_sig: BlockBuilder<ExprBuilder<BlockBuilder<ItemImplMethodBuilder<ItemImplBuilder<Identity>>>>>)
-> ItemImplBuilder<Identity> where T: ToAstType<T> + Narrow<u32>

Generates a setter for a given offset+index.

Given the offset+index type, it will optimize read-before-write unless fully byte aligned. Also generates documentation for the setter.

fn build_setters(&self,
off: &IoRegOffsetInfo,
seg: &IoRegSegmentInfo,
prev: ItemImplBuilder<Identity>)
-> ItemImplBuilder<Identity>

Loops the offset+info blocks for this segment and generates all setters.

fn lookup_const_val<'b>(&'b self,
name: &str,
seg: &'b IoRegSegmentInfo)
-> Result<&StaticValue, String>

Gets the StaticValue from the constants defined from this ioreg, with an error if it does not exist.

NOTE: this looks up the constants using the scoped name, but scope prefix is done for you by the function.

fn build_peripheral_init_fns(&self,
impl_block: ItemImplBuilder<Identity>)
-> ItemImplBuilder<Identity>

entry for generating the impl block for the mcu.