Struct dynamo::ioreg::common::IoRegOffsetIndexInfo

pub struct IoRegOffsetIndexInfo {
    pub offset: u8,
    pub width: u8,
    pub span: Span,
}

Represents the offset from the segment, and the index into the register of a register-portion.

This is used to define where functions can access and gives mask/optimization info to the Builder.

Fields
offset: u8 width: u8 span: Span
Methods

impl IoRegOffsetIndexInfo

fn width_is_byte_aligned(&self) -> bool

fn offset_is_byte_aligned(&self) -> bool

fn is_fully_byte_aligned(&self) -> bool

fn offset_in_bytes(&self) -> u32