Struct dynamo::ioreg::common::IoRegSegmentInfo

pub struct IoRegSegmentInfo {
    pub name: String,
    pub address: u32,
    pub reg_width: RegisterWidth,
    pub access_perms: RegisterPermissions,
    pub const_vals: BTreeMap<String, StaticValue>,
    pub offsets: Vec<IoRegOffsetInfo>,
    pub span: Span,
}

Internal representation of a Segment.

Segments are typically what the mcu will expose as a register, but we want to group all related segments into one struct. Segments contain multiple offsets which themselves contain the functions that act on that "register".

Fields
name: String address: u32 reg_width: RegisterWidth access_perms: RegisterPermissions const_vals: BTreeMap<String, StaticValue> offsets: Vec<IoRegOffsetInfo> span: Span
Methods

impl IoRegSegmentInfo

fn push_offset(&mut self, off: IoRegOffsetInfo)

Add an offset to the segment.

fn can_read(&self) -> bool

Truthy calue on whether the segment can be read from.

fn can_write(&self) -> bool

Truthy calue on whether the segment can be written to.