Struct dynamo::mcu::builder::Builder

pub struct Builder<'a> {
    // some fields omitted
}

Builds the generated AST from the parsed mcu!() block.

Consumes the parser to later set syntax errors.

Methods

impl<'a> Builder<'a>

fn new(mcu: McuInfo, parser: Parser<'a>, verbose: bool) -> Builder

Consumes the parsed common::McuInfo descriptor and parser to construct a Builder.

fn build(&self) -> Vec<P<Item>>

Generates the AST from the descriptor block.

This includes: 1. externs 2, the MCU struct 3. the MCU impl (including the ::new() function) 4. the interrupt array

If the no_static keyword is omitted, the following are also generated: 1. the Sync trait as locking is left to the implementor 2. a static version of the mcu exported as MCU

fn build_struct(&self) -> P<Item>

Builds the MCU structure definition including derived traits, doc comments, and doc sources.

Only the definition, traits, and doc comments are generated directly. Everything else uses helpers.

fn build_sync_impl(&self) -> P<Item>

Generate the Sync trait for our type.

This is onyl done if a static version of the mcu is generated.

fn build_new_struct(&self) -> P<Expr>

Generates the structure initializer used in both the ::new() function and to initialize the static version (if generated).

fn build_static_instantiation(&self) -> P<Item>

Generates the static MCU: MyMcu = MyMcu{...}; statement if specified in the expansion.

fn build_externs(&self) -> P<Item>

Generates the extern "C" { ... } block for any defined externs in the expansion.

fn build_impl(&self) -> P<Item>

Entry for generating the impl block for the mcu.

fn build_copy_data(&self, impl_block: ItemImplBuilder<Identity>) -> ItemImplBuilder<Identity>

Builds the ::copy_data_section() function for the MCU definition.

This function will copy the range [src_begin, src_end] to the data destination location in memory.

fn build_interrupts(&self) -> P<Item>

Builds the interrupts table and associates the #[link_section = ".some_location"] attribute with it.

The array generated is of type [Option<fn()>; NUM_INTERRUPTS] and defaults to None for all interrupts.

fn build_stack_and_entry_ptrs(&self) -> Vec<P<Item>>

Generates the static values for the stack_base and code_entry pointers.

These need to be places at specific locations, so a static value with a #[link_section] attr is used.