Module dynamo::mcu::common

Structs

ActionInfo

BssInfo
Internal structure representing the parsed bss block.

DataInfo
Internal structure representing the parsed data block.

HeapInfo
Internal structure representing the parsed heap block.

InitInfo

InterruptsInfo
Internal structure representing the parsed interrupts block.

McuInfo
Internal structure for the builder which describes the parsed MCU block.

NvicInfo

PeripheralInfo
Internal structure describing a peripheral to be used by the builder to add it to the MCU.

RangeInfo
Structure representing either an index (single number) or a range (1..3).

StackInfo
Internal structure representing the parsed stack block.