Enum dynamo::parser::StaticValue

pub enum StaticValue {
    Int(i32, String, Span),
    Uint(u32, String, Span),
    Float(f32, InternedString, String, Span),
    Str(String, String, Span),
    Ident(String, Ident, Span),
    Path(Path, Span),
    Error(String, Span),
}

Tuple type representing a static value parsed from the input.

Certain value types can only be used in certain places/actions. Each type carries the Span for which it was defined to properly place errors.

Variants
Int(i32, String, Span)Uint(u32, String, Span)Float(f32, InternedString, String, Span)Str(String, String, Span)Ident(String, Ident, Span)Path(Path, Span)Error(String, Span)
Methods

impl StaticValue

fn default_uint() -> StaticValue

fn to_string(&self) -> String

fn get_span(&self) -> Span