Struct dynamo::parser::CommonParser

pub struct CommonParser<'a> {
    pub parser: Parser<'a>,
    pub begin_segment: Span,
    // some fields omitted
}

The common parser functions used across this crate.

Essentially a thin wrapper around syntax::parse::parser::Parser, with helpers added for our shared syntax.

Fields
parser: Parser<'a> begin_segment: Span
Methods

impl<'a> CommonParser<'a>

fn from(cx: &mut ExtCtxt<'a>, tree: &[TokenTree]) -> CommonParser<'a>

Create a parser from the parsing context and token tree we wish to parse.

fn set_err(&mut self, err: &str)

Sets a syntax error on the span the parser's cursor is currently on.

fn set_err_last(&mut self, err: &str)

Sets a syntax error on the span before the parser's current token cursor.

fn set_fatal_err(&mut self, err: &str)

Sets a fatal error on the parser's current token, and emits the diagnostic.

fn set_fatal_err_last(&mut self, err: &str)

Sets a fatal error on the parser's previous token, and emits the diagnostic.

fn set_segment_err(&mut self, err: &str)

Sets a syntax error on the span representing the current segment.

Note: this will be moving to the ioreg parser shortly as this is generic, but missed in the split.

fn expect_ident_value(&mut self, expect: &str)

Expects the current token to be an ast::Ident, and that the ident's name is equal to the given value.

If the values are not equal, a fatal error is set.

fn expect_semi(&mut self) -> bool

Expect the current token to be a semicolon

If the current token is not the expect token, a fatal error is set.

fn expect_equal(&mut self) -> bool

Expect the current token to be an equal sign ('=')

If the current token is not the expect token, a fatal error is set.

fn expect_fat_arrow(&mut self) -> bool

Expect the current token to be a fat arrow ('=>')

If the current token is not the expect token, a fatal error is set.

fn expect_colon(&mut self) -> bool

Expect the current token to be a colon

If the current token is not the expect token, a fatal error is set.

fn expect_comma(&mut self) -> bool

Expect the current token to be a comma

If the current token is not the expect token, a fatal error is set.

fn expect_open_paren(&mut self) -> bool

Expect the current token to be an opening paren ('('))

If the current token is not the expect token, a fatal error is set.

fn expect_close_paren(&mut self) -> bool

Expect the current token to be a closing paren (')')

If the current token is not the expect token, a fatal error is set.

fn expect_open_curly(&mut self) -> bool

Expect the current token to be an opening curly brace ('{')

If the current token is not the expect token, a fatal error is set.

fn expect_close_curly(&mut self) -> bool

Expect the current token to be a closing curly brace ('}')

If the current token is not the expect token, a fatal error is set.

fn expect_open_bracket(&mut self) -> bool

Expect the current token to be an opening bracket ('[')

If the current token is not the expect token, a fatal error is set.

fn expect_close_bracket(&mut self) -> bool

Expect the current token to be a closing bracket (']')

If the current token is not the expect token, a fatal error is set.

fn get_ident(&mut self) -> Ident

Get the parser's current token as an ast::Ident.

If the current token is not an identifier, a non-fatal syntax error is set and a dummy is returned.

fn get_type_path(&mut self) -> Path

Get the parser's current token(s) as an ast::Path.

If the current token is not a path, a non-fatal syntax error is set and a dummy is returned.

fn get_literal(&mut self) -> Result<Lit, String>

Get the parser's current token as an ast::Lit.

If the current token is not a literal, a non-fatal syntax error is set and an error is returned.

fn parse_uint<T: HasMinMax<T> + From<T>>(&mut self) -> u64 where u64: From<T>

Parse the current token (i.e. literal) as a unsigned integer.

This function is essentially a sanity wrapper around checked_parse_uint(). We cannot genericize the return value, so it is returned as a u64 which should then be cast appropriately.

Note: there should be changes to the return value coming shortly, if Rust supports it.

fn checked_parse_uint<T: HasMinMax<T> + From<T>>(&mut self) -> Result<u64, String> where u64: From<T>

Reads the parser's current token as an unsigned numeric literal, asserting it fits into T::max_value().

If the current token is not a numeric literal or the parsed value does not fit within T::max_value() then an error is returned describing as such.

fn parse_constant_literal(&mut self, name: &String) -> StaticValue

Parses the parser's current token as a literal, and converts it to our StaticValue type.

Note: Numeric literals are parsed as u32, and as such, must fit within u32::max_value(). Note: Signed integers are not yet supported.

There are multiple error cases here, depending on the type of literal.

fn parse_ident_string(&mut self) -> String

Parses the current token as an ident, and returns the string value of the identifier.

If the token is not an ident, a syntax error is placed and a dummy value returned.

fn parse_doc_sources(&mut self, prefix: &String, into: &mut Vec<String>)

Parses an entire doc_srcs => [ ... ] block into the given vector

Note: trailing commas are not allowed.

fn parse_constants_block<T>(&mut self, prefix: &String, into: &mut BTreeMap<String, StaticValue>, validate: fn(&T, &StaticValue, &mut BTreeMap<String, StaticValue>) -> Result<(), String>, validate_ctx: &T)

Parse an entire constants => { ... }; block into the given hash map.

If any constant fails to parse, a syntax error is placed and parsing is aborted. This can lead to further, somewhat confusing, syntax errors but it follows the Rust style.

NOTE: this function takes a validator function as an argument in case any validation needs to happen. This can be a nop.