Struct dynamo::mcu::builder::Builder

pub struct Builder<'a> {
    // some fields omitted
}

Builds the generated AST from the parsed mcu!() block.

Consumes the parser to later set syntax errors.

Methods

impl<'a> Builder<'a>

fn new(mcu: McuInfo, parser: Parser<'a>, verbose: bool) -> Builder

Consumes the parsed common::McuInfo descriptor and parser to construct a Builder.

fn build(&self) -> Vec<P<Item>>

Generates the AST from the consumed descriptor block.

fn build_struct(&self) -> P<Item>

Builds the MCU structure definition including doc comments, impl functions, etc.

fn build_new_struct(&self) -> P<Expr>

Generates the structure initializer used in both the ::new() function and to initialize the static version (if generated).

fn build_externs(&self) -> P<Item>

Generates the extern "C" { ... } block for any defined externs in the expansion.

fn build_mcu_init_fn(&self, impl_block: ItemImplBuilder<Identity>) -> ItemImplBuilder<Identity>

entry for generating the impl block for the mcu.

fn build_impl(&self) -> P<Item>

Entry for generating the impl block for the mcu.

fn build_copy_data(&self, impl_block: ItemImplBuilder<Identity>) -> ItemImplBuilder<Identity>

Builds the ::copy_data_section() function for the MCU definition.

This function will copy the range [src_begin, src_end] to the data destination location in memory.

fn build_null_bss(&self, impl_block: ItemImplBuilder<Identity>) -> ItemImplBuilder<Identity>

Builds the ::null_bss() function for the MCU definition.

fn build_interrupts(&self) -> Vec<P<Item>>

Builds the interrupts table and associates the #[link_section = ".some_location"] attribute with it.

The type is actually a struct containing the stack pointer as a u32 and an array of ISRs. The array generated is of type [core::option::Option<fn()>; NUM_INTERRUPTS-1] and defaults to None (0u32) for all interrupts. Making this #[repr(C)] structure keeps type safety while achieving the packed table structure we need.

fn build_nvic_ty(&self) -> P<Item>

Generates the NVIC type we will add to the generated MCU.

This structure contains the function needed to enable/disable ints, set prios, etc. It will look and function just like an ioreg peripheral, but generated here instead.

fn build_nvic_instance(&self) -> P<Expr>

Generates an ast::Expr that describes the instantiation of the NVIC.

This is used for NVIC::new() as well as the MCU instantiation.

fn build_nvic_impl(&self) -> P<Item>

Generates the functions that will exist on the NVIC handler, but not the NVIC trait.

fn build_nvic_trait_impl(&self) -> P<Item>

fn build_nvic_trait_fns(&self, impl_block: ItemImplBuilder<Identity>) -> ItemImplBuilder<Identity>

Generates the functions that will exist on the NVIC handler.

fn build_init(&self) -> P<Item>

Build the init() function that instantiates the mcu.

The generated code is as follows:

pub fn init() {
    let mcu = FooMcu::new();
    mcu.init();
    unsafe { main(mcu); }
}